dutch_National_Flag_Algorithm
Desc: Given an array nums with n objects colored red, white, or blue, sort them in-place so that objects of the same color are adjacent, with the colors in the order red, white, and blue.
Here, we will use the integers 0, 1, and 2 to represent the color red, white, and blue respectively.

This algorithm helps to arrange 0s,1s,2s in an array in just O(n) time complexity and O(1) Extra space. This is the most optimal solution.