Create Super Market Billing.cpp
This is a project I Have done previously. This may through some errors as I have different compiler to run . There might be issue with header files too. But overall it is a genuine code. Please do some editing to run this code in your IDE. 
This is also contribution of Hacktober fest 2020.
Please merge this also in your repo
Thank you